fix : eslint error

+++ src/Repository/UserRepository.ts
@@ -1,45 +1,25 @@
 import pool from '../config/database';
-import { UserRow } from 'src/interface/both';
+import { type UserRow } from 'src/interface/both';
 import { type FieldPacket } from 'mysql2';
 
 export const createUser = async (uid: string) => {
-  try {
-    const connection = await pool.getConnection();
-    await connection.execute('INSERT INTO user (uid) VALUES (?)', [uid]);
-    connection.release();
-  } catch (error) {
-    throw error;
-  }
+  const connection = await pool.getConnection();
+  await connection.execute('INSERT INTO user (uid) VALUES (?)', [uid]);
+  connection.release();
 };
 
 export const readUser = async (uid: string): Promise<UserRow[]> => {
-  try {
-    const connection = await pool.getConnection();
-
-    const [user]: [UserRow[], FieldPacket[]] = await connection.execute('SELECT * FROM user WHERE uid = ?', [uid]);
-    connection.release();
-    return user;
-  } catch (error) {
-    throw error;
-  }
+  const connection = await pool.getConnection();
+  const [user]: [UserRow[], FieldPacket[]] = await connection.execute('SELECT * FROM user WHERE uid = ?', [uid]);
+  connection.release();
+  return user;
 };
 
 export const removeUser = async (uid: string): Promise<void> => {
-  try {
-    const connect = await pool.getConnection();
+  const connect = await pool.getConnection();
+  const userDeleteQuery: string = 'DELETE FROM user WHERE uid = ?';
 
-    const userDeleteQuery: string = 
-    `
-    DELETE FROM user
-    WHERE uid = ?
-    `;
-
-    await connect.execute<UserRow[]>(userDeleteQuery, [
-      uid
-    ]);
-  } catch (error) {
-    throw error;
-  }
-}
+  await connect.execute<UserRow[]>(userDeleteQuery, [uid]);
+};
 
 export default { createUser, readUser, removeUser };
